requirements.txt
pandas
numpy
scikit-learn
matplotlib
yfinance
#cdp-sdk
ta
plotly
kat
backtesting
seaborn
statsmodels.api
statsmodels.tsa.api
